% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{cols}
\alias{col_character}
\alias{col_date}
\alias{col_datetime}
\alias{col_double}
\alias{col_factor}
\alias{col_guess}
\alias{col_integer}
\alias{col_logical}
\alias{col_number}
\alias{col_skip}
\alias{col_time}
\alias{cols_condense}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{readr}{\code{\link[readr:parse_atomic]{col_character}}, \code{\link[readr:parse_datetime]{col_date}}, \code{\link[readr:parse_datetime]{col_datetime}}, \code{\link[readr:parse_atomic]{col_double}}, \code{\link[readr:parse_factor]{col_factor}}, \code{\link[readr:parse_guess]{col_guess}}, \code{\link[readr:parse_atomic]{col_integer}}, \code{\link[readr:parse_atomic]{col_logical}}, \code{\link[readr:parse_number]{col_number}}, \code{\link[readr]{col_skip}}, \code{\link[readr:parse_datetime]{col_time}}, \code{\link[readr]{cols}}, \code{\link[readr:spec]{cols_condense}}}
}}

